Add Tox configuration
Can use this for code quality checks ("tox -e check, etc.), running
tests, and configuring tools. When running tests, it runs each test
in a isolated environment created using virtualenv. This makes it
trivial to test all the versions of Python we support, provided they
are currently installed (pyenv is useful for this on Linux).